Fix drsyms on Win x64.
- Fix a size_t -> DWORD truncation warning.
- We cannot walk the stack reliably on Win x64, so we give up on that
  part of the test for now, and symbolize only the first frame.

SVN-Revision: 1112

+++ ext/drsyms/drsyms_windows.c
@@ -657,7 +657,8 @@ drsym_demangle_symbol(char *dst OUT, size_t dst_sz, const char *mangled,
         undec_flags = UNDNAME_NAME_ONLY;
     }
 
-    len = (size_t)UnDecorateSymbolName(mangled, dst, dst_sz, undec_flags);
+    len = (size_t)UnDecorateSymbolName(mangled, dst, (DWORD)dst_sz,
+                                       undec_flags);
 
     /* The truncation behavior is not documented, but testing shows dbghelp
      * truncates and returns the number of characters written, not how many it
